requirements.txt
asgiref==3.7.2
certifi==2022.6.15
cffi==1.16.0
charset-normalizer==2.1.0
cryptography==41.0.4
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
gunicorn==21.2.0
idna==3.3
netaddr==0.8.0
oauthlib==3.2.2
packaging==23.1
paypal==1.2.5
Pillow==10.0.0
psutil==5.9.5
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
requests==2.28.1
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
stripe==6.6.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.11
whitenoise==6.5.0
